move simulation_settings to generic_tech folder
