+++ gdsfactory/technology/__init__.py
@@ -1,10 +1,7 @@
 from gdsfactory.technology.layer_map import lyp_to_dataclass
 from gdsfactory.technology.layer_stack import LayerLevel, LayerStack
 from gdsfactory.technology.layer_views import LayerView, LayerViews
-from gdsfactory.technology.simulation_settings import (
-    SIMULATION_SETTINGS_LUMERICAL_FDTD,
-    SimulationSettingsLumericalFdtd,
-)
+
 
 __all__ = [
     "LayerView",
@@ -12,6 +9,4 @@ __all__ = [
     "LayerLevel",
     "LayerStack",
     "lyp_to_dataclass",
-    "SimulationSettingsLumericalFdtd",
-    "SIMULATION_SETTINGS_LUMERICAL_FDTD",
 ]
